Add shared lib target, rename static lib (fix #55)
Code using multiple copies of the static libraries in one process can
cause problems due to multiple copies of the protobuffer libraries not
interacting correctly. Change default target library to be a shared
library, add static variant as _static, and use object libraries to
optimize the building of the shared and _pic static libraries.