Make version number defaults 0 to catch mismatches
This fixes issue #92: Making the current version default values will
result in sending code not sending the fields, and the receiving code
will substitute its default values, which will be the receiver's version
and not the sender's version, as intended. The only backward compatible
change is to change the default values to 0, and require senders to
fill out the field prior to sending out the message.  Since proto2/3
do not directly support defining constants, this currently entails
manually setting the proper values. In a future commit this might be
changed to use option extensions to provide the current version to
the user code as option values.