Add initial draft of sensor input configuration message
This is a first draft of the message used to configure the sensor input
between the model and the simulation environment. The mechanics of the
message interchange are determined by the model packaging.

It should be noted that the sensor input configuration mechanism is
separate from any mechanisms that are used to configure the sensor
model itself, which currently are packaging dependent.