Add host vehicle data as sensor input data
This changeset introduces HostVehicleData which comprises all vehicle
internal data that is available to the sensors (or in the future other
parts of OSI) as additional input. This includes e.g. GPS-sensor derived
data as well as e.g. engine ECU data, etc.  For now this only includes
GPS-sensor location data with error estimation.