Port per-major-release cmake install from v2.2.1
The cmake file from current master branch is reused and adjusted to
allow several OSI versions on the same computer.

The main idea is to install into a directory that takes the version
into account.

This draft uses the major version described inside the version file
to configure the installation path.