Implementing the "make install" for the OSI 2.
The cmake file from current master branch is reuse and adjusted to allow several OSI versions on the same computer.

The main idea is to install the into a directory that takes the version into account.

This draft uses the major version described inside the version file to configure the installation path.

Correcting copy error.