Use subTest to show all failing files
When one file is failing, due to the
fact that all tests contain for loops,
the other failures are swallowed because python
stops running the test on the first assertion.
With subTest one can introduce a subcontext which
allows showing all of the failed files as well a
diagnostic as to which file failed.