Properly describe lanes on intersections
Currently, OSI does not describe lanes on intersections. This is
problematic for agent models, which need the lane information to drive
properly.

Change OSI to describe lanes on intersections as normal lanes.
Intersection lanes can be distinguished from non-intersection lanes by
the new `intersection` field.

This way of describing intersections is equivalent to how intersections are
described in OpenDRIVE and NDS. Thus, it becomes much easier to convert
from OpenDRIVE or NDS to OSI, and no information is lost anymore.

This is an incompatible change, so it is 4.0 material (note: could be made
compatible if needed, by only deprecating TYPE_INTERSECTION).

TODO:
- Update the intersection images