Add road type based on OpenDRIVE to logical lanes
The road type is useful since traffic participant models might want to
influence their behavior based on the type of road that they are driving
on, or make decisions in terms of navigation based on that information.
It is also helpful to display a road in a navigation system, as
typically different road types are displayed differently.

Signed-off-by: Daniel Mata <Daniel.Mata@partner.bmw.de>